:bug: :sparkles: :pencil: :construction: :arrow_up: :recycle: :heavy_plus_sign: :pencil2: :alien: :bento: and more
I did a ton of stuff
Update deps
Update everything to match the new REI version
Code cleanup
Refinery! (WIP)
Add a bunch of assets that I have been ignoring
A bunch of stuff I probably forgot

+++ src/main/java/com/hrznstudio/galacticraft/world/dimension/MarsDimension.java
@@ -63,7 +63,7 @@ public class MarsDimension extends Dimension implements CelestialBody {
         f3 = f3 * (f2 * 0.94F + 0.06F);
         f4 = f4 * (f2 * 0.94F + 0.06F);
         f5 = f5 * (f2 * 0.91F + 0.09F);
-        return new Vec3d((double) f3, (double) f4, (double) f5);
+        return new Vec3d(f3, f4, f5);
     }
 
     @Override
